@@ -2,6 +2,5 @@
 
 namespace Common.Types;
 
-public class ValidationError : NewType<ValidationError, string> {
-  public ValidationError(string e) : base(e) { }
+public class ValidationError(string e) : NewType<ValidationError, string>(e) {
 }
